Use KiCAD to Create Schematic
This adds a new KiCAD v5 project to the repository, containing a new
board schematic with a custom symbol for the voltage regulator, as well
a footprint for our eventual SMD inductor.

The new schematic includes a voltage regulation circuit for running off
of battery power, a reset button, a crystal oscillator circuit,
decoupling capacitors on the chip's power pins & ISP/FTDI headers.