Fix Footprint Size for C2, Fix Silkscreen for L1
Create a new CP_Elec footprint using the datasheet for C2, swapping out
the footprint in the PCB schematic.

Fix the Silkscreen layer for the L1 footprint, ensuring it does not
cross the pads.

Shift the SW2 footprint up so that it's courtyard does not overlap the
courtyard of J4.